refactor: move handleSesssionCreatedEvent fn to utils folder
