+++ src/utils/handleSessionCreatedEvent.ts
